refactor(admin): separate settings and changing log controllers
